Add app mounted event > GTM > Intercom boot
Sending an app mounted event when the app mounts to GTM.
Triggering Intercom.boot() in GTM when this happens to make it
easier to toggle it on and off without re-deploying the app.

Intercom.update() is triggered on the pageview event, which looks
for new messages that should be displayed to the current user.