Handle message with no responsePath
Fixing: https://sentry.io/hedvig/hedvig-app/issues/475492285/activity/

Currently the backend always sets responsePath to /response on messages,
this is used when POSTing a response from the chat.

We should hardcode this on the front end until we need this functionality
to make debugging and tracability easier in the codebase.