Hide asset tracker tab
The current incarnation of the asset tracker seems to cause confusion for
users around if they actually need to add their stuff to have it covered.

There is no legal need to register stuff to have it covered. It's a nice
to have for users so our hypothesis is that the current version just
confuses and creates more work to find receipts etc.

We might want to restore this functionality at some point but until we do
we should not have dead code hanging in the repo, this makes debugging harder.

We might restore it as a way to add extra coverage for stuff above 50k
so it might be completely redesigned by this point.