Attempt to fix scaling issue for peril images
There's an issue in prod where the peril images are
not scaled scaled properly in the home screen carousel.

This problem does not exist in test with current prod code,
so attempting to change the scaling from center to contain
(which shoud have the same effect) and re-publishing to see if
there was some issue with the last expo bundle and publish.