Fix issue where you got no initial messages
To initiate a converstation (login or signup) the previous logic POSTed
to /chat/start and separately chat/conversation view GETs /messages

If the /chat/start (or /chat/login) POST fails or completes after GET
/messages no conversation will be started and no messages are returned in
the chat view.

If the request fails the user is stuck on the chat view without
any messages, even restarting the app woudln't fix it as the /chat/start
POST happened from the marketing screen.

We should not rely on order of async requests and should never expect a
request to have happened on a previous view for the current view to get
the data it needs.

Sending an intent param when GETing /messages (intent=signup|login) and
remove calls to /chat/start and /chat/login.

The backend uses the intent to start the right conversation if no existing messages exist.
Defaults to signup if no intent is given.

The intent param is stored in the state and persisted in AsyncStorage using
redux-persist.

Using redux-persist seemed like the least crazy way of making sure it's
persisted and the app gated until the store has been rehydrated from the
async cache.

Also we'll be able to fix the issue where you reset the converstation from
login but get the onboarding/signup conversation instead.

- Removed unused keyboard event listeners
- Renamed all rotues and consolidated all nav setup (preparing caching nav state)
- Added Loader spinner to Chat, Offer, Dashboard and Profile view if data isn't properly loaded
- Clear AsyncStorage when app crashes so users are not left with potentially corrupt caches
- Rename Carousel to Perils
- Remove generic ChatModal and use explicit route Offer instead
- Fix Navigation reset on initial redirect and logout where the key has to be null to reset root nav