Register campaing on all deep link opens
Whenever a deep link is openend and it has campaign params from Branch
send these to the backend and register the campaign.

Previously we sent the first install params which are only set once on the
same device even if you re-install the app.
So you could use a campaign link
later and not get attributed on this campaign.
This also made it impossible to test campaigns.